Cleanup generate_messages.rs
